Solutions/Lab 4 (Chapter 3.1)/requirements.txt
# Requirements cho Lab 4 - Wavelet Image Similarity
# Python 3.8+

# Xử lý ảnh và wavelet
numpy>=1.21.0
Pillow>=9.0.0
PyWavelets>=1.4.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Machine learning metrics
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Pandas cho xử lý dữ liệu
pandas>=1.4.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.0.0
